LayoutTree: Factor out DropCliff.
