+++ src/components/LayoutTree.tsx
@@ -35,8 +35,7 @@ import isRoot from '../util/isRoot'
 import parentOf from '../util/parentOf'
 import parseLet from '../util/parseLet'
 import safeRefMerge from '../util/safeRefMerge'
-import unroot from '../util/unroot'
-import DropEnd from './DropEnd'
+import DropCliff from './DropCliff'
 import VirtualThought from './VirtualThought'
 
 /** 1st Pass: A thought with rendering information after the tree has been linearized. */
@@ -714,41 +713,15 @@ const LayoutTree = () => {
                   cliff < 0 &&
                   // do not render hidden cliffs
                   // rough autofocus estimate
-                  autofocusDepth - depth < 2 &&
-                  Array(-cliff)
-                    .fill(0)
-                    .map((x, i) => {
-                      const pathEnd = -(cliff + i) < path.length ? (path.slice(0, cliff + i) as Path) : HOME_PATH
-                      const cliffDepth = unroot(pathEnd).length
-
-                      // After table col2, shift the DropEnd left by the width of col1.
-                      // This correctly positions the drop target for dropping after the table view.
-                      // Otherwise it would be too far to the right.
-                      const dropEndMarginLeft =
-                        isTableCol2 && cliffDepth - depth < 0 ? treeThoughtsPositioned[index - 1].width || 0 : 0
-
-                      return (
-                        <div
-                          key={'DropEnd-' + head(pathEnd)}
-                          className='z-index-subthoughts-drop-end'
-                          style={{
-                            position: 'relative',
-                            top: '-0.2em',
-                            left: `calc(${cliffDepth - depth}em - ${dropEndMarginLeft}px + ${isTouch ? -1 : 1}px)`,
-                            transition: 'left 0.15s ease-out',
-                          }}
-                        >
-                          <DropEnd
-                            depth={pathEnd.length}
-                            path={pathEnd}
-                            // Extend the click area of the drop target when there is nothing below.
-                            // The last visible drop-end will always be a dimmed thought at distance 1 (an uncle).
-                            // Dimmed thoughts at distance 0 should not be extended, as they are dimmed siblings and sibling descendants that have thoughts below
-                            // last={!nextChildId}
-                          />
-                        </div>
-                      )
-                    })}
+                  autofocusDepth - depth < 2 && (
+                    <DropCliff
+                      cliff={cliff}
+                      depth={depth}
+                      isTableCol2={isTableCol2}
+                      prevWidth={treeThoughtsPositioned[index - 1].width}
+                      path={path}
+                    />
+                  )}
               </div>
             )
           },
